+++ apps/io-session-manager-internal/src/services/__tests__/session-service.test.ts
@@ -53,7 +53,6 @@ import {
   toConflictError,
   toGenericError,
   toNotFoundError,
-  toNotFoundError,
 } from "../../utils/errors";
 import { aNotReleasedData } from "../../__mocks__/table-client.mock";
 import { LoginTypeEnum } from "../../types/fast-login";