+++ apps/io-session-manager-oi/src/domain/ports/outbound/ausiliar-data.port.ts
@@ -1,0 +1,16 @@
+import { GenericError, NotFoundError } from "@pagopa/hexagonal-core";
+import { HealthCheckOutboundPort } from "@pagopa/io-auth-n-identity-domain";
+import { Result } from "neverthrow";
+
+import { LoginAusiliarData } from "../../value-objects/login.vo.js";
+
+export interface AusiliarDataPort extends HealthCheckOutboundPort {
+  readonly save: (
+    id: string,
+    obj: LoginAusiliarData,
+  ) => Promise<Result<undefined, GenericError>>;
+
+  readonly retrieve: (
+    id: string,
+  ) => Promise<Result<LoginAusiliarData, GenericError | NotFoundError>>;
+}