+++ apps/io-session-manager-oi/src/domain/value-objects/oidc.vo.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+export const OidcConfigurationEnvSchema = z.union([
+  z.literal("UAT"),
+  z.literal("PROD"),
+]);
+export type OidcConfigurationEnv = z.infer<typeof OidcConfigurationEnvSchema>;