+++ infra/resources/modules/session_manager/ca.tf
@@ -17,8 +17,12 @@ module "sm_ca" {
 
   secrets = [
     {
-      name                = "REDIS_PASSWORD"
-      key_vault_secret_id = azurerm_key_vault_secret.sm_redis_access_key.versionless_id
+      name                = "ONEID_PROD_CLIENT_SECRET"
+      key_vault_secret_id = azurerm_key_vault_secret.sm_oneid_prod_client_secret.versionless_id
+    },
+    {
+      name                = "ONEID_UAT_CLIENT_SECRET"
+      key_vault_secret_id = azurerm_key_vault_secret.sm_oneid_uat_client_secret.versionless_id
     },
   ]
 
@@ -28,7 +32,7 @@ module "sm_ca" {
       name  = "${var.prefix}-${var.domain}-${local.app_name}"
 
       app_settings = local.app_settings
-      secret_names = ["REDIS_PASSWORD"]
+      secret_names = ["ONEID_PROD_CLIENT_SECRET", "ONEID_UAT_CLIENT_SECRET"]
 
       liveness_probe = {
         path = "/api/auth/v2/health/liveness"