+++ infra/resources/modules/session_manager/kv_secrets.tf
@@ -18,8 +18,18 @@ resource "azurerm_key_vault_secret" "sm_io_profile_api_key" {
   tags = var.tags
 }
 
-resource "azurerm_key_vault_secret" "sm_redis_access_key" {
-  name         = "sm-redis-access-key"
+resource "azurerm_key_vault_secret" "sm_oneid_prod_client_secret" {
+  name         = "sm-oneid-prod-client-secret"
+  key_vault_id = var.key_vault.id
+
+  value_wo         = ""
+  value_wo_version = 1
+
+  tags = var.tags
+}
+
+resource "azurerm_key_vault_secret" "sm_oneid_uat_client_secret" {
+  name         = "sm-oneid-uat-client-secret"
   key_vault_id = var.key_vault.id
 
   value_wo         = ""